@@ -1,16 +1,18 @@
-import React from 'react'
-import ReactDOM from 'react-dom'
-import {Provider} from 'react-redux'
-import {Router} from 'react-router-dom'
-import history from './history'
-import store from './store'
-import App from './App'
+import React from "react";
+import ReactDOM from "react-dom/client";
+import { Provider } from "react-redux";
+import { BrowserRouter as Router } from "react-router-dom";
+import history from "./history";
+import store from "./store";
+import App from "./App";
 
-ReactDOM.render(
+const container = document.getElementById("app");
+const root = ReactDOM.createRoot(container);
+
+root.render(
   <Provider store={store}>
-    <Router history={history}>
+    <Router>
       <App />
     </Router>
-  </Provider>,
-  document.getElementById('app')
-)
+  </Provider>
+);
